refactor(svg_render): Split widget builders into per-widget modules
