+++ custom_components/eink_dashboard/svg_render.py
@@ -20,9 +20,8 @@ import functools
 import json
 from collections.abc import Callable
 from dataclasses import fields as dc_fields
-from datetime import datetime
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, NamedTuple, cast
+from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:
     from .render import WidgetMetrics
@@ -40,8 +39,6 @@ from .const import (
     COLOR_WHITE,
     DEFAULT_CARD_STYLE,
     DEFAULT_ROW_H,
-    FONT_SIZE_WEATHER,
-    PADDING,
     WidgetType,
 )
 
@@ -635,1371 +632,6 @@ def _auto_row_height(
     return svg_h
 
 
-def _build_separator_context(
-    widget: Widget,
-    config: DisplayConfig,
-) -> dict[str, object]:
-    """Build Jinja2 template context for the separator widget.
-
-    Computes the SVG viewport dimensions and the bounding rectangle
-    for the separator element.  Both ``"line"`` and ``"bar"`` styles
-    are represented as a single ``<rect>`` whose width and height are
-    pre-computed here so the template needs no conditionals.
-
-    The ``"bar"`` style widens to 10 px on 2-level displays
-    (``grayscale_levels <= 2``) so the dithered dot pattern reads
-    clearly as a separator.
-
-    Args:
-        widget: Widget config dict.  Recognised keys:
-            ``direction`` (``"horizontal"`` | ``"vertical"``,
-            default ``"horizontal"``),
-            ``style`` (``"line"`` | ``"bar"``,
-            default ``"line"``),
-            ``length`` (explicit pixel length; omit for full
-            span), ``x`` (default ``PADDING``),
-            ``y`` (default 0).
-        config: Display config with ``width``, ``height``, and
-            optional ``grayscale_levels`` (default 16).
-
-    Returns:
-        Dict consumed by ``separator.svg.j2``: ``w``, ``h``,
-        ``bar_w``, ``bar_h``, ``fill``.
-    """
-    from .render import color_to_hex
-
-    x = widget.get("x", PADDING)
-    y = widget.get("y", 0)
-    direction = widget.get("direction", "horizontal")
-    style = widget.get("style", "line")
-    grayscale_levels = config.get("grayscale_levels", 16)
-
-    svg_w = _widget_dim(widget, "w", config["width"] - x)
-    svg_h = _widget_dim(widget, "h", config["height"] - y)
-
-    if style == "bar":
-        color: int = COLOR_GRAY
-        # Widen bar on 2-level displays so the dithered dot
-        # pattern reads clearly as a separator.
-        thickness = 10 if grayscale_levels <= 2 else 6
-    else:
-        color = COLOR_BLACK
-        thickness = 2
-
-    # Default span: viewport dimension minus one PADDING unit,
-    # matching the PIL formula config[dim] - PADDING - pos.
-    explicit_length: int | None = widget.get("length")
-    if explicit_length is not None:
-        length = explicit_length
-    elif direction == "vertical":
-        length = svg_h - PADDING
-    else:
-        length = svg_w - PADDING
-
-    fill = color_to_hex(color)
-
-    if direction == "vertical":
-        bar_w: int = thickness
-        bar_h: int = length
-    else:
-        bar_w = length
-        bar_h = thickness
-
-    return {
-        "w": svg_w,
-        "h": svg_h,
-        "bar_w": bar_w,
-        "bar_h": bar_h,
-        "fill": fill,
-        **_color_context(),
-    }
-
-
-# Weather base geometry at scale=1.0
-# (font_size == FONT_SIZE_WEATHER == 32).  Each value is multiplied
-# by `scale` in _build_weather_context() to adapt to the configured
-# font_size.  These constants are weather-specific and have no
-# equivalent in WidgetMetrics; padding and divider thickness are
-# derived from _compute_metrics() directly inside the builder.
-#
-# _WX_ROW_H must be defined first; other derived constants reference it.
-_WX_ROW_H = 48  # 48, not DEFAULT_ROW_H (56): matches original PIL proportions
-_WX_NATURAL_W = 380  # natural card width
-_WX_ICON = 80  # condition icon diameter
-_WX_FONT_XL = 64  # temperature font size (bold)
-_WX_FONT_SM = 16  # hi/lo, detail, and forecast font
-_WX_FONT_XS = 14  # precipitation text font
-_WX_ICON_R_PAD = 16  # gap: condition icon → temp text
-_WX_DETAIL_GAP = 2  # vertical gap above detail row
-_WX_DETAIL_ICON_H = 20  # detail icon height
-_WX_ICON_GAP = 4  # gap: detail icon → its text
-_WX_SEP_GAP = 8  # gap above/below separator line
-_WX_FC_ZONE_H = 88  # forecast zone height
-_WX_PRECIP_H = _WX_FONT_SM  # line height matches font
-_WX_FC_ICON = 32  # forecast day icon diameter
-_WX_FC_ICON_CY = 34  # forecast icon centre Y offset
-_WX_FC_HI_Y = 52  # forecast hi-temp text Y offset
-_WX_FC_LO_Y = 70  # forecast lo-temp text Y offset
-_WX_FC_PRECIP_Y = _WX_FC_ZONE_H  # precip text at zone bottom
-_WX_MIN_FC_COLS = 5  # minimum forecast column count
-_WX_LO_Y_FRAC = 0.4  # lo temp Y as fraction of temp_h
-_WX_PRECIP_Y_FRAC = 0.72  # precip text Y as fraction of temp_h
-
-_DETAIL_ICON_MAP: dict[str, str] = {
-    "humidity": "wi-humidity",
-    "barometer": "wi-barometer",
-    "wind": "wi-strong-wind",
-    "cloud": "wi-cloud",
-}
-
-
-def _cap_weather_font_xl(
-    font_xl_size: int,
-    font_xl: Any,
-    font_sm: Any,
-    font_xs: Any,
-    temp_text: str,
-    avail: int,
-    today_hi: str,
-    today_lo: str,
-    today_precip: str,
-) -> int:
-    """Return font_xl capped so the temperature text fits in avail px.
-
-    Measures the widest hi/lo/precipitation string to determine how far
-    that column protrudes leftward from its right anchor, then reduces
-    font_xl proportionally if the temperature text would overlap it.
-
-    Args:
-        font_xl_size: Nominal xl font size in pixels.
-        font_xl: PIL font loaded at font_xl_size (for text measurement).
-        font_sm: PIL font for hi/lo text measurement.
-        font_xs: PIL font for precipitation text measurement.
-        temp_text: Formatted temperature string (e.g. "13.8°C").
-        avail: Pixel budget between temp_x and the hi/lo column.
-        today_hi: High-temperature string (may be empty).
-        today_lo: Low-temperature string (may be empty).
-        today_precip: Precipitation string (may be empty).
-
-    Returns:
-        Capped font size; equals font_xl_size when text already fits.
-    """
-    hilo_w = 0
-    if today_hi:
-        hilo_w = max(hilo_w, round(font_sm.getlength(today_hi)))
-    if today_lo:
-        hilo_w = max(hilo_w, round(font_sm.getlength(today_lo)))
-    if today_precip:
-        hilo_w = max(hilo_w, round(font_xs.getlength(today_precip)))
-    budget = avail - hilo_w
-    temp_w = round(font_xl.getlength(temp_text))
-    if budget > 0 and temp_w > budget:
-        return round(font_xl_size * budget / temp_w)
-    return font_xl_size
-
-
-def _build_weather_context(
-    widget: Widget,
-    config: DisplayConfig,
-) -> dict[str, object]:
-    """Build Jinja2 template context for the weather widget.
-
-    Pre-computes every position and icon SVG string so the
-    Jinja2 template contains no layout logic.
-
-    Args:
-        widget: Widget config dict.  Recognised keys:
-            ``entity``, ``x``, ``y``, ``w``, ``font_size``,
-            ``forecast_days``, ``card_style``.
-        config: Display config with ``width``, ``height``,
-            ``states``, ``grayscale_levels``.
-
-    Returns:
-        Template context dict consumed by ``weather.svg.j2``.
-        Returns ``{"w": …, "h": …, "has_state": False}`` when
-        the entity is absent from ``states``.
-    """
-    # Lazy imports avoid circular dependency: render.py imports
-    # svg_render.py at module level; if svg_render.py imported
-    # render.py at module level the initialisation would fail.
-    from .render import (
-        _DAY_ABBREV,
-        _compute_metrics,
-        _fmt_temp,
-        _load_font,
-    )
-
-    entity_id = widget.get("entity", "")
-    state = config.get("states", {}).get(entity_id)
-    x = widget.get("x", PADDING)
-    svg_w = _widget_dim(widget, "w", config["width"] - x)
-
-    if state is None:
-        svg_h = _widget_dim(widget, "h", config["height"] - widget.get("y", 0))
-        return {"w": svg_w, "h": svg_h, "has_state": False, **_color_context()}
-
-    font_size = widget.get("font_size", FONT_SIZE_WEATHER)
-    forecast_days = widget.get("forecast_days", 5)
-    card_style = widget.get("card_style", DEFAULT_CARD_STYLE)
-    grayscale_levels = config.get("grayscale_levels", 16)
-
-    scale = font_size / FONT_SIZE_WEATHER
-
-    # Card width: use explicit w or natural width capped to canvas.
-    w_override = widget.get("w")
-    if w_override is not None:
-        card_w = w_override
-    else:
-        card_w = min(round(_WX_NATURAL_W * scale), svg_w)
-        # Clip SVG to content width so the editor resize box
-        # matches the rendered content, not the full canvas.
-        svg_w = _widget_dim(widget, "w", card_w)
-
-    # PIL fonts for text measurement only — never used for
-    # drawing.  Affects: temp_h, temp_bbox (getbbox) and
-    # text_w_i (getlength) below.
-    # Bold is used for font_xl because the temperature text renders
-    # with font-weight="bold" in the SVG template.
-    font_xl = _load_font(round(_WX_FONT_XL * scale), bold=True)
-    font_sm = _load_font(round(_WX_FONT_SM * scale))
-
-    # Entity attributes.
-    condition = state.get("state", "")
-    attrs = state.get("attributes", {})
-    temp = attrs.get("temperature", "--")
-    temp_unit = attrs.get("temperature_unit", "°C")
-    humidity = attrs.get("humidity")
-    wind = attrs.get("wind_speed")
-    wind_unit = attrs.get("wind_speed_unit", "km/h")
-    pressure = attrs.get("pressure")
-    pressure_unit = attrs.get("pressure_unit", "hPa")
-    cloud_coverage = attrs.get("cloud_coverage")
-    forecast = attrs.get("forecast", [])
-
-    # Card metrics — 48 at scale=1 gives card-level metrics
-    # (padding~10, radius~10) matching the original PIL layout.
-    m = _compute_metrics(round(_WX_ROW_H * scale))
-
-    # Sizing constants, all proportional to scale.
-    icon_size = round(_WX_ICON * scale)
-    pad = m.padding
-    icon_right_pad = round(_WX_ICON_R_PAD * scale)
-    detail_gap = round(_WX_DETAIL_GAP * scale)
-    detail_icon_h = round(_WX_DETAIL_ICON_H * scale)
-    icon_gap = round(_WX_ICON_GAP * scale)
-    sep_gap = round(_WX_SEP_GAP * scale)
-    sep_thickness = m.divider
-    forecast_zone_h = round(_WX_FC_ZONE_H * scale)
-    precip_text_h = round(_WX_PRECIP_H * scale)
-
-    # Measure temperature text height (PIL) for height estimation.
-    nf = config.get("number_format", "language")
-    lang = config.get("language", "en")
-    temp_text = f"{_fmt_temp(temp, nf, lang)}{temp_unit}"
-    temp_bbox = font_xl.getbbox(temp_text)
-    temp_h = temp_bbox[3] - temp_bbox[1]
-
-    top_pad = m.padding
-
-    # Total card height, matching PIL's formula exactly.
-    row1_h = top_pad + max(icon_size, temp_h)
-    detail_h = detail_gap + detail_icon_h
-    has_forecast = bool(forecast) and forecast_days > 0
-    if has_forecast:
-        forecast_section_h = (
-            sep_gap + sep_thickness + sep_gap + forecast_zone_h + precip_text_h
-        )
-    else:
-        forecast_section_h = pad
-    total_h = row1_h + detail_h + forecast_section_h + pad
-    # Default to content height so the SVG is no taller than its
-    # rendered content.  Without this, the editor resize box spans
-    # the full remaining canvas when no explicit h is configured.
-    svg_h = _widget_dim(widget, "h", total_h)
-
-    # Content insets — "border" and "left_bar" use the shared
-    # helper; "none" applies its own pad so that the weather card
-    # content never touches the viewport edge.
-    if card_style == "none":
-        content_left = pad
-        content_w = card_w - 2 * pad
-        bar_width = 0
-    else:
-        x_off, r_inset, bar_width = _card_insets(
-            m, card_style, grayscale_levels
-        )
-        content_left = x_off
-        content_w = card_w - x_off - r_inset
-
-    content_top = top_pad
-
-    # Row 1: condition icon + temperature + today hi/lo/precip.
-    icon_cy = content_top + icon_size // 2
-    icon_x = content_left
-    icon_y = content_top
-    temp_x = content_left + icon_size + icon_right_pad
-    # dominant-baseline="central" in template — centres em-square
-    # on icon_cy, matching PIL's visible-ink centering within a
-    # few pixels.
-    temp_y = icon_cy
-
-    # vis_top: top of the visible temperature glyph, used as
-    # anchor for the stacked hi/lo/precip text block.
-    vis_top = icon_cy - temp_h // 2
-    hilo_right = content_left + content_w - pad
-
-    today_hi = ""
-    today_lo = ""
-    today_precip = ""
-    lo_y = vis_top + round(temp_h * _WX_LO_Y_FRAC)
-    precip_y = vis_top + round(temp_h * _WX_PRECIP_Y_FRAC)
-    precip_unit_fc = attrs.get("precipitation_unit", "mm")
-    if forecast:
-        today = forecast[0]
-        hi_val = today.get("temperature")
-        lo_val = today.get("templow")
-        p_val = today.get("precipitation")
-        if hi_val is not None:
-            today_hi = f"{_fmt_temp(hi_val, nf, lang)}°"
-        if lo_val is not None:
-            today_lo = f"{_fmt_temp(lo_val, nf, lang)}°"
-        if p_val is not None:
-            today_precip = f"{_fmt(str(p_val), config)}{precip_unit_fc}"
-
-    # Cap font_xl so temp text doesn't overlap the hi/lo column.
-    font_xl_size = _cap_weather_font_xl(
-        round(_WX_FONT_XL * scale),
-        font_xl,
-        font_sm,
-        _load_font(round(_WX_FONT_XS * scale)),
-        temp_text,
-        hilo_right - temp_x - pad,
-        today_hi,
-        today_lo,
-        today_precip,
-    )
-
-    # row1_bottom mirrors PIL's max() between icon bottom and
-    # the bottom of the temperature glyph.
-    temp_y_pil = icon_cy - temp_bbox[1] - temp_h // 2
-    row1_bottom = max(
-        content_top + icon_size,
-        temp_y_pil + temp_bbox[3],
-    )
-
-    # Condition icon SVG.
-    try:
-        cond_icon_svg: markupsafe.Markup | str = _weather_svg_filter(
-            condition, icon_size
-        )
-    except (KeyError, FileNotFoundError):
-        cond_icon_svg = ""
-
-    # Detail row: icon + text pairs for weather attributes.
-    detail_y = row1_bottom + detail_gap
-    raw_details: list[tuple[str, str]] = []
-    if humidity is not None:
-        raw_details.append(("humidity", f"{_fmt(str(humidity), config)}%"))
-    if pressure is not None:
-        raw_details.append(
-            (
-                "barometer",
-                f"{_fmt(str(round(pressure)), config)}{pressure_unit}",
-            )
-        )
-    if wind is not None:
-        raw_details.append(
-            (
-                "wind",
-                f"{_fmt(str(round(wind)), config)}{wind_unit}",
-            )
-        )
-    if cloud_coverage is not None:
-        raw_details.append(
-            (
-                "cloud",
-                f"{_fmt(str(cloud_coverage), config)}%",
-            )
-        )
-
-    detail_cols = max(len(raw_details), 1)
-    col_w_detail = content_w // detail_cols
-    detail_items: list[dict[str, object]] = []
-
-    for i, (icon_name, text) in enumerate(raw_details):
-        col_cx = content_left + col_w_detail * i + col_w_detail // 2
-        text_w_i = round(font_sm.getlength(text))
-        svg_filename = _DETAIL_ICON_MAP.get(icon_name, "")
-        # Wrap in Markup so Jinja2 emits the SVG verbatim.  All
-        # icon strings added to the context must be Markup instances.
-        detail_icon_svg: markupsafe.Markup | str = ""
-        if svg_filename:
-            detail_path = (_ICONS_DIR / f"{svg_filename}.svg").resolve()
-            try:
-                detail_paths = _load_svg_paths(detail_path)
-                detail_icon_svg = markupsafe.Markup(
-                    _build_inline_svg(detail_paths, detail_icon_h, "0 0 30 30")
-                )
-            except FileNotFoundError:
-                pass
-        has_detail_icon = bool(detail_icon_svg)
-        icon_w = detail_icon_h + icon_gap if has_detail_icon else 0
-        item_w = icon_w + text_w_i
-        item_x = col_cx - item_w // 2
-        detail_items.append(
-            {
-                "icon_svg": detail_icon_svg,
-                "icon_x": item_x,
-                "icon_y": detail_y,
-                "text_x": (
-                    item_x + detail_icon_h + icon_gap
-                    if has_detail_icon
-                    else item_x
-                ),
-                "text_y": detail_y + detail_icon_h // 2,
-                "text": text,
-            }
-        )
-
-    detail_bottom = detail_y + detail_icon_h
-
-    # Forecast grid.
-    forecast_entries: list[dict[str, object]] = []
-    sep_x1 = 0
-    sep_x2 = 0
-    sep_y = 0
-
-    if has_forecast:
-        forecast_cols = max(forecast_days, _WX_MIN_FC_COLS)
-        col_width = content_w // forecast_cols
-        content_width = forecast_cols * col_width
-        separator_y = detail_bottom + sep_gap
-        sep_x1 = content_left
-        sep_x2 = content_left + content_width
-        sep_y = separator_y
-        # sep_thickness accounts for the separator line height so
-        # forecast content starts below the stroke bottom, matching
-        # the sep_thickness term in forecast_section_h.
-        forecast_y = separator_y + sep_thickness + sep_gap
-        fc_icon_size = round(_WX_FC_ICON * scale)
-
-        if forecast_days >= forecast_cols:
-            col_positions = list(range(forecast_days))
-        elif forecast_days <= 1:
-            col_positions = [forecast_cols // 2]
-        else:
-            col_positions = [
-                round(i * (forecast_cols - 1) / (forecast_days - 1))
-                for i in range(forecast_days)
-            ]
-
-        for idx, day in enumerate(forecast[:forecast_days]):
-            col_i = col_positions[idx]
-            cx = content_left + col_width * col_i + col_width // 2
-            dt_str = day.get("datetime")
-            if dt_str:
-                day_label = _DAY_ABBREV[
-                    datetime.fromisoformat(dt_str).weekday()
-                ]
-            else:
-                day_label = ""
-
-            day_condition = day.get("condition", "")
-            try:
-                fc_icon_svg: markupsafe.Markup | str = _weather_svg_filter(
-                    day_condition, fc_icon_size
-                )
-            except (KeyError, FileNotFoundError):
-                fc_icon_svg = ""
-
-            fc_hi_val = day.get("temperature", "")
-            fc_lo_val = day.get("templow", "")
-            fc_hi = (
-                f"{_fmt_temp(fc_hi_val, nf, lang)}°" if fc_hi_val != "" else ""
-            )
-            fc_lo = (
-                f"{_fmt_temp(fc_lo_val, nf, lang)}°" if fc_lo_val != "" else ""
-            )
-            fc_p = day.get("precipitation")
-            fc_precip = (
-                f"{_fmt(str(fc_p), config)}{precip_unit_fc}"
-                if fc_p is not None and fc_p > 0
-                else ""
-            )
-            icon_cy_fc = forecast_y + round(_WX_FC_ICON_CY * scale)
-            forecast_entries.append(
-                {
-                    "cx": cx,
-                    "label": day_label,
-                    "label_y": forecast_y,
-                    "icon_svg": fc_icon_svg,
-                    "icon_x": cx - fc_icon_size // 2,
-                    "icon_y": icon_cy_fc - fc_icon_size // 2,
-                    "hi": fc_hi,
-                    "hi_y": forecast_y + round(_WX_FC_HI_Y * scale),
-                    "lo": fc_lo,
-                    "lo_y": forecast_y + round(_WX_FC_LO_Y * scale),
-                    "precip": fc_precip,
-                    "precip_y": forecast_y + round(_WX_FC_PRECIP_Y * scale),
-                }
-            )
-
-    return {
-        "w": svg_w,
-        "h": svg_h,
-        "has_state": True,
-        "card_w": card_w,
-        "total_h": total_h,
-        "card_style": card_style,
-        **_metrics_context(m),
-        "bar_width": bar_width,
-        "icon_svg": cond_icon_svg,
-        "icon_x": icon_x,
-        "icon_y": icon_y,
-        "icon_size": icon_size,
-        "temp_text": temp_text,
-        "temp_x": temp_x,
-        "temp_y": temp_y,
-        "font_xl": font_xl_size,
-        "font_sm": round(_WX_FONT_SM * scale),
-        # font_xs is template-only; no PIL measurement needed.
-        "font_xs": round(_WX_FONT_XS * scale),
-        "hilo_right": hilo_right,
-        "hi_text": today_hi,
-        "hi_y": vis_top,
-        "lo_text": today_lo,
-        "lo_y": lo_y,
-        "precip_text": today_precip,
-        "precip_y": precip_y,
-        "detail_items": detail_items,
-        "has_forecast": has_forecast,
-        "sep_x1": sep_x1,
-        "sep_x2": sep_x2,
-        "sep_y": sep_y,
-        "sep_thickness": sep_thickness,
-        "forecast_entries": forecast_entries,
-        **_color_context(),
-    }
-
-
-def _build_device_battery_context(
-    widget: Widget,
-    config: DisplayConfig,
-) -> dict[str, object]:
-    """Build Jinja2 template context for the device_battery widget.
-
-    Supports two layouts via the ``layout`` parameter:
-
-    - ``"icon"`` (default): compact battery outline with percentage
-      label, sized via h-based proportional ratios.  At ``h=40``
-      the standard geometry applies (bw=30, bh=14, nub_w=3,
-      nub_h=8).
-    - ``"chip"``: pill-shaped chip with a proportional fill bar
-      and percentage label, sized via ``h``.
-
-    An optional ``card_style`` parameter wraps the content in a
-    card frame.  Content insets are pre-computed from the card
-    style so the template receives final absolute positions.
-
-    Args:
-        widget: Widget config dict.  Recognised keys: ``x``,
-            ``y``, ``w``, ``h`` (default 40), ``layout``,
-            ``card_style``, ``color``.
-        config: Display config with ``device_battery_level``
-            (int 0–100) and ``grayscale_levels``.
-
-    Returns:
-        Template context dict consumed by
-        ``device_battery.svg.j2``.  Returns
-        ``{"w": …, "h": …, "has_level": False}`` when
-        ``device_battery_level`` is absent from config.
-    """
-    from .render import _compute_metrics, _load_font, color_to_hex
-
-    x = widget.get("x", PADDING)
-    svg_w = _widget_dim(widget, "w", config["width"] - x)
-    # h: raw geometry reference for proportional calculations.
-    # svg_h: clamped SVG viewport height (minimum 1 via _widget_dim).
-    h: int = widget.get("h", 40)
-    svg_h = _widget_dim(widget, "h", 40)
-
-    level = config.get("device_battery_level")
-    if level is None:
-        return {"w": svg_w, "h": svg_h, "has_level": False, **_color_context()}
-
-    pct = max(0, min(100, int(level)))
-    layout = widget.get("layout", "icon")
-    card_style = widget.get("card_style", DEFAULT_CARD_STYLE)
-    grayscale_levels = config.get("grayscale_levels", 16)
-    color: int = widget.get("color", COLOR_BLACK)
-    # Force black below 20% for visual emphasis.
-    if pct < 20:
-        color = COLOR_BLACK
-    color_hex = color_to_hex(color)
-
-    label = f"{pct}%"
-    m = _compute_metrics(svg_h)
-
-    x_off, r_inset, bar_width = _card_insets(m, card_style, grayscale_levels)
-
-    if layout == "chip":
-        # svg_w is the available width (explicit w or remaining
-        # canvas); used as upper bound for chip content before
-        # svg_w is narrowed to actual content extent below.
-        content_w = svg_w - x_off - r_inset
-        # pad, gap, and font_sz match the chip macro ratios;
-        # bar_w_nat and bar_h are battery-specific geometry.
-        pad = h * 18 // 100
-        gap = h * 14 // 100
-        bar_w_nat = h * 120 // 100
-        bar_h = h * 36 // 100
-        bar_border = max(1, m.border // 2)
-        font_sz = max(10, h * 46 // 100)
-        # PIL font for text measurement only — resvg does not
-        # expose text metrics, so widths are pre-computed here.
-        font = _load_font(font_sz)
-        text_w = round(font.getlength(label))
-
-        chip_w = min(pad + bar_w_nat + gap + text_w + pad, content_w)
-        # Reflow bar to fit within a capped chip.
-        bar_w = max(0, chip_w - pad - gap - text_w - pad)
-
-        chip_radius = h // 2
-        bar_y = (h - bar_h) // 2
-        fill_w = int((bar_w - 2) * pct / 100) if bar_w > 2 else 0
-
-        # Clip SVG width to chip content plus card insets so the
-        # editor resize box matches the rendered content.
-        w_override = widget.get("w")
-        if w_override is not None:
-            svg_w = max(1, w_override)
-        else:
-            svg_w = max(1, x_off + chip_w + r_inset)
-
-        return {
-            "w": svg_w,
-            "h": svg_h,
-            "has_level": True,
-            "layout": "chip",
-            "card_h": svg_h,
-            "card_style": card_style,
-            **_metrics_context(m),
-            **_color_context(),
-            "bar_width": bar_width,
-            "color_hex": color_hex,
-            "label": label,
-            "font_sz": font_sz,
-            "chip_x": x_off,
-            "chip_w": chip_w,
-            "chip_radius": chip_radius,
-            "bar_abs_x": x_off + pad,
-            "bar_y": bar_y,
-            "bar_w": bar_w,
-            "bar_h": bar_h,
-            "bar_border": bar_border,
-            "fill_abs_x": x_off + pad + 1,
-            "fill_y": bar_y + 1,
-            "fill_w": fill_w,
-            "fill_h": max(0, bar_h - 2),
-            "text_abs_x": x_off + pad + bar_w + gap,
-            "text_y": h // 2,
-        }
-
-    # Icon layout: compact battery outline with proportional fill bar.
-    # Ratios chosen so that h=40 produces the standard geometry
-    # (body_w=30, body_h=14, nub_w=3, nub_h=8).
-    body_w = round(h * 0.75)
-    body_h = round(h * 0.35)
-    nub_w = round(h * 0.075)
-    nub_h = round(h * 0.20)
-    nub_gap = max(1, round(h * 0.025))
-    gap = round(h * 0.10)
-    font_sz = max(10, round(h * 0.60))
-    font = _load_font(font_sz)
-    # 'la' (left-ascender) is the default anchor for
-    # FreeTypeFont.getbbox(), centring the battery body on the
-    # visible text glyph ink rather than the full EM square.
-    # Clamp to 0 so negative ascender values never push the rect
-    # above the SVG canvas.
-    bbox = font.getbbox(label)
-    text_h = bbox[3] - bbox[1]
-    icon_y = max(0, bbox[1] + (text_h - body_h) // 2)
-    nub_y = icon_y + (body_h - nub_h) // 2
-    fill_w = int((body_w - 2) * pct / 100)
-
-    # Clip SVG width to icon+text content plus card insets so
-    # the editor resize box matches the rendered content.
-    w_override = widget.get("w")
-    if w_override is not None:
-        svg_w = max(1, w_override)
-    else:
-        svg_w = max(
-            1,
-            x_off + body_w + nub_gap + nub_w + gap + round(bbox[2]) + r_inset,
-        )
-
-    return {
-        "w": svg_w,
-        "h": svg_h,
-        "has_level": True,
-        "layout": "icon",
-        "card_h": svg_h,
-        "card_style": card_style,
-        **_metrics_context(m),
-        **_color_context(),
-        "bar_width": bar_width,
-        "color_hex": color_hex,
-        "label": label,
-        "font_sz": font_sz,
-        "body_x": x_off,
-        "icon_y": icon_y,
-        "body_w": body_w,
-        "body_h": body_h,
-        "nub_abs_x": x_off + body_w + nub_gap,
-        "nub_y": nub_y,
-        "nub_w": nub_w,
-        "nub_h": nub_h,
-        "fill_abs_x": x_off + 1,
-        "icon_fill_y": icon_y + 1,
-        "fill_w": fill_w,
-        "icon_fill_h": max(0, body_h - 2),
-        "text_abs_x": x_off + body_w + nub_gap + nub_w + gap,
-        "text_svg_y": icon_y + body_h // 2,
-    }
-
-
-def _build_waste_schedule_context(
-    widget: Widget,
-    config: DisplayConfig,
-) -> dict[str, object]:
-    """Build Jinja2 template context for the waste_schedule widget.
-
-    Renders waste collection entries as card rows with urgency
-    styling.  Each entry maps an entity attribute (ISO date string
-    or integer day offset) to a short label and colour scheme:
-
-    - ``days <= 1``: black-filled icon circle; date text is black
-      when ``days == 0`` (today) and gray otherwise.
-    - ``days >= 2``: outlined icon circle (white fill, black
-      stroke); date text is gray.
-
-    Two layouts are supported via the ``layout`` parameter:
-
-    - ``"list"`` (default): one row per visible entry, each with a
-      gray horizontal divider between rows; date is shown as a
-      right-aligned ``value`` string.
-    - ``"card"``: only the most urgent entry (lowest day count)
-      using the full widget height; date is shown as a
-      ``secondary`` line below the primary label.
-
-    By default only entries within the 0–3 day range are rendered.
-    When ``show_all`` is ``True``, entries with any future date are
-    included regardless of distance.  Entries with missing,
-    unparseable, or past dates are always silently skipped.
-
-    Args:
-        widget: Widget config dict.  Recognised keys: ``entity``
-            (entity ID whose attributes hold the dates),
-            ``entries`` (list of ``{"attribute": …, "label": …}``
-            dicts), ``layout`` (``"list"`` or ``"card"``),
-            ``show_all`` (``bool``, default ``False``),
-            ``card_style``, ``title``, ``x``, ``w``, ``h``.
-        config: Display config with ``states`` (entity ID →
-            state dict) and ``grayscale_levels``.
-
-    Returns:
-        Template context dict consumed by
-        ``waste_schedule.svg.j2``.  Returns
-        ``{"w": …, "h": …, "has_rows": False}`` when the entity
-        is absent from states, ``entries`` is empty, or no entries
-        fall within the visible day range.
-    """
-    from .render import (
-        _compute_metrics,
-        _format_relative_date,
-        _get_today,
-        _parse_days_until,
-        color_to_hex,
-    )
-
-    x = widget.get("x", PADDING)
-    svg_w = _widget_dim(widget, "w", config["width"] - x)
-
-    entity_id: str = widget.get("entity", "")
-    entries: list[dict[str, str]] = widget.get("entries", [])
-    layout: str = widget.get("layout", "list")
-    card_style = widget.get("card_style", DEFAULT_CARD_STYLE)
-    title: str = widget.get("title", "")
-    show_all: bool = bool(widget.get("show_all", False))
-    states = config.get("states", {})
-    grayscale_levels = config.get("grayscale_levels", 16)
-
-    empty_ctx: dict[str, object] = {
-        "w": svg_w,
-        "h": _widget_dim(widget, "h", DEFAULT_ROW_H),
-        "has_rows": False,
-        **_color_context(),
-    }
-    if not entity_id or not entries:
-        return empty_ctx
-
-    entity_state = states.get(entity_id)
-    if entity_state is None:
-        return empty_ctx
-
-    attrs = entity_state.get("attributes", {})
-
-    # Resolve visible entries before computing height so the
-    # auto-sizing fallback knows how many rows to accommodate.
-    # Config order is preserved — equal-day entries keep the
-    # order the user configured them, matching the PIL renderer.
-    # Use _get_today() from render.py so tests can patch
-    # "custom_components.eink_dashboard.render.date" to control
-    # the date returned here.
-    today = _get_today()
-    visible: list[tuple[str, str, int]] = []
-    for entry in entries:
-        attr_key = entry.get("attribute", "")
-        label = entry.get("label") or attr_key
-        raw = str(attrs.get(attr_key, ""))
-        if not raw:
-            continue
-        days = _parse_days_until(raw, today)
-        if days is None or days < 0:
-            continue
-        if not show_all and days > 3:
-            continue
-        visible.append((label, raw, days))
-
-    if not visible:
-        return empty_ctx
-
-    if layout == "card":
-        # Most urgent entry only (stable sort: equal days keep
-        # config order), displayed at full widget height.
-        visible.sort(key=lambda e: e[2])
-        visible = [visible[0]]
-
-    # Auto-size: default to content height so the widget is no
-    # taller than its rendered rows.  Explicit "h" in the widget
-    # config overrides this, preserving backward compatibility.
-    num_display_rows = len(visible)
-    svg_h = _widget_dim(widget, "h", _auto_row_height(title, num_display_rows))
-    title_font_sz, content_y, content_h = _title_layout(title, svg_h)
-    # card layout always trims visible to one entry above, so
-    # num_display_rows == 1 and the division is equivalent to content_h.
-    row_h = content_h // num_display_rows
-
-    m = _compute_metrics(row_h)
-    icon_stroke_w = m.border * 3 if grayscale_levels <= 2 else m.border
-    divider_stroke_w = m.divider * 3 if grayscale_levels <= 2 else m.divider
-    x_off, r_inset, bar_width = _card_insets(m, card_style, grayscale_levels)
-    lpad = m.padding if x_off == 0 else 0
-    rpad = m.padding if r_inset == 0 else 0
-
-    # Build the trash-can icon SVG once; all entries share the
-    # same icon, only the circle fill/outline differs.
-    icon_sz = m.icon_inner
-    icon_svg = _mdi_svg_filter("trash-can", icon_sz)
-
-    rows: list[dict[str, object]] = []
-    for i, (label, raw, days) in enumerate(visible):
-        date_str = _format_relative_date(days, raw)
-        # days == 0 (today): black date text for urgency.
-        # days >= 1: gray date text.
-        date_fill = (
-            color_to_hex(COLOR_BLACK)
-            if days == 0
-            else color_to_hex(COLOR_GRAY)
-        )
-        use_outline = days >= 2
-        # Outlined circles ignore icon_fill (macro uses white
-        # fill + black stroke), but we still pass it correctly.
-        icon_fill = (
-            color_to_hex(COLOR_BLACK)
-            if days <= 1
-            else color_to_hex(COLOR_GRAY)
-        )
-        rows.append(
-            {
-                "y": content_y + i * row_h,
-                "primary": label,
-                # Card layout: date as secondary (below primary).
-                # List layout: date as right-aligned value.
-                "secondary": date_str if layout == "card" else "",
-                "value": "" if layout == "card" else date_str,
-                "icon_svg": icon_svg,
-                "icon_outline": use_outline,
-                "icon_fill": icon_fill,
-                "secondary_fill": date_fill,
-                "value_fill": date_fill,
-                "letter": "",
-            }
-        )
-
-    return {
-        "w": svg_w,
-        "h": svg_h,
-        "has_rows": True,
-        "title": title,
-        "title_font_sz": title_font_sz,
-        "content_y": content_y,
-        "content_h": content_h,
-        "card_style": card_style,
-        "bar_width": bar_width,
-        **_metrics_context(m),
-        **_color_context(),
-        "row_h": row_h,
-        "rows": rows,
-        "x_off": x_off,
-        "r_inset": r_inset,
-        "lpad": lpad,
-        "rpad": rpad,
-        "icon_stroke_w": icon_stroke_w,
-        "divider_stroke_w": divider_stroke_w,
-    }
-
-
-def _build_tile_context(
-    widget: Widget,
-    config: DisplayConfig,
-) -> dict[str, object]:
-    """Build Jinja2 template context for the tile widget.
-
-    Single-entity card with icon circle on the left, entity name
-    as primary text, and optional state text as secondary.  Mirrors
-    HA's Tile card (``hui-tile-card.ts``).
-
-    Icon circle style is controlled by the ``icon_style`` config
-    field:
-
-    - ``"filled"`` — gray circle background, white icon (default
-      for active entities).
-    - ``"outlined"`` — white fill, black stroke circle, black icon
-      (default for inactive entities).
-    - ``"none"`` — no circle; icon is rendered in black.
-
-    When ``icon_style`` is absent the style is chosen automatically:
-    2-level displays always use ``"outlined"``; multi-level displays
-    use ``"filled"`` for active states (per ``_ACTIVE_STATES``) and
-    ``"outlined"`` otherwise.
-
-    Args:
-        widget: Widget config dict.  Recognised keys:
-            ``entity`` (required entity ID),
-            ``name`` (display name override),
-            ``icon`` (MDI icon override, e.g. ``"mdi:lightbulb"``),
-            ``hide_state`` (suppress secondary text),
-            ``state_content`` (attribute name or list; first element
-            used when a list is provided),
-            ``icon_style`` (``"filled"`` / ``"outlined"`` /
-            ``"none"``), ``card_style``,
-            ``x``, ``w``, ``h``.
-        config: Display config with ``width``, ``height``,
-            ``states``, and ``grayscale_levels``.
-
-    Returns:
-        Template context dict consumed by ``tile.svg.j2``.
-        Returns ``{"w": …, "h": …, "has_entity": False}`` when
-        the entity is absent from ``states``.  Full context
-        includes: widget dimensions (``w``, ``h``), layout
-        geometry (``row_h``, ``x_off``, ``r_inset``, ``lpad``,
-        ``rpad``), metrics fields (``m_*`` prefix), color hex
-        strings (``hex_*`` prefix), text content (``primary``,
-        ``secondary``), icon data (``icon_svg``, ``letter``),
-        and icon style flags (``icon_fill``, ``icon_outline``,
-        ``icon_no_circle``).  The ``card_row`` ``value`` key is
-        intentionally omitted — tiles have no right-aligned
-        value text; the macro defaults to ``""``.
-    """
-    from .render import (
-        _compute_metrics,
-        _device_class_icon,
-        color_to_hex,
-    )
-
-    x = widget.get("x", PADDING)
-    svg_w = _widget_dim(widget, "w", config["width"] - x)
-    entity_id: str = widget.get("entity", "")
-    name_override = widget.get("name")
-    icon_override = widget.get("icon")
-    hide_state: bool = widget.get("hide_state", False)
-    state_content = widget.get("state_content")
-    icon_style = widget.get("icon_style")
-    card_style = widget.get("card_style", DEFAULT_CARD_STYLE)
-    states = config.get("states", {})
-    grayscale_levels = config.get("grayscale_levels", 16)
-
-    state = states.get(entity_id) if entity_id else None
-    if state is None:
-        return {
-            "w": svg_w,
-            "h": _widget_dim(widget, "h", DEFAULT_ROW_H),
-            "has_entity": False,
-            **_color_context(),
-        }
-
-    # Single row: row_h equals svg_h.  Kept as a separate
-    # variable so card_row receives the same parameter name
-    # as in multi-row widgets (e.g. waste_schedule).
-    svg_h = _widget_dim(widget, "h", _auto_row_height("", 1))
-    row_h = svg_h
-    m = _compute_metrics(row_h)
-    x_off, r_inset, bar_width = _card_insets(m, card_style, grayscale_levels)
-    # Zero lpad/rpad when card_container already insets that side.
-    lpad = m.padding if x_off == 0 else 0
-    rpad = m.padding if r_inset == 0 else 0
-
-    attrs = state.get("attributes", {})
-    domain = entity_id.split(".")[0] if entity_id else ""
-    state_val = state.get("state", "")
-
-    # Primary text: name override or entity friendly_name.
-    primary: str = (
-        str(name_override)
-        if name_override is not None
-        else attrs.get("friendly_name", entity_id)
-    )
-
-    # Secondary text: attribute, default state+unit, or hidden.
-    if hide_state:
-        secondary: str = ""
-    elif state_content is not None:
-        sc: str = (
-            state_content[0]
-            if isinstance(state_content, list) and state_content
-            else (state_content if not isinstance(state_content, list) else "")
-        )
-        secondary = _fmt(str(attrs.get(sc, "")), config) if sc else ""
-    else:
-        unit = attrs.get("unit_of_measurement", "")
-        fmtd = _fmt(state_val, config)
-        secondary = f"{fmtd}{unit}" if unit else fmtd
-
-    # Icon: explicit override → letter fallback when not found.
-    # No override → device_class → letter fallback.
-    # When an explicit icon is requested and its file is absent,
-    # skip device_class lookup so the user's intent is respected.
-    icon_svg: markupsafe.Markup | str = ""
-    if icon_override is not None:
-        icon_name = str(icon_override)
-        if icon_name.startswith("mdi:"):
-            icon_name = icon_name[4:]
-        with contextlib.suppress(FileNotFoundError):
-            icon_svg = _mdi_svg_filter(icon_name, m.icon_inner)
-    else:
-        resolved_name = _device_class_icon(attrs, state_val, domain)
-        if resolved_name is None:
-            raw = attrs.get("icon", "")
-            if raw.startswith("mdi:"):
-                resolved_name = raw[4:]
-        if resolved_name:
-            with contextlib.suppress(FileNotFoundError):
-                icon_svg = _mdi_svg_filter(resolved_name, m.icon_inner)
-
-    letter = ""
-    if not icon_svg:
-        friendly = attrs.get("friendly_name", entity_id)
-        letter = friendly[:1].upper() if friendly else ""
-
-    # Icon style: explicit config overrides auto-switching.
-    is_active = state_val in _ACTIVE_STATES
-    if icon_style is None:
-        # 2-level displays use outlined for maximum contrast.
-        # Multi-level displays switch by entity state.
-        if grayscale_levels <= 2:
-            resolved_style = "outlined"
-        elif is_active:
-            resolved_style = "filled"
-        else:
-            resolved_style = "outlined"
-    else:
-        resolved_style = str(icon_style)
-
-    icon_outline = resolved_style == "outlined"
-    icon_no_circle = resolved_style == "none"
-    # Filled style always uses gray; state is conveyed by
-    # icon_style (filled vs outlined), not fill colour.
-    icon_fill = color_to_hex(COLOR_GRAY)
-    # Widen the outline stroke on 2-level displays to avoid dithering.
-    icon_stroke_w = m.border * 3 if grayscale_levels <= 2 else m.border
-
-    return {
-        "w": svg_w,
-        "h": svg_h,
-        "has_entity": True,
-        "card_style": card_style,
-        "bar_width": bar_width,
-        **_metrics_context(m),
-        **_color_context(),
-        "row_h": row_h,
-        "x_off": x_off,
-        "r_inset": r_inset,
-        "lpad": lpad,
-        "rpad": rpad,
-        "primary": primary,
-        "secondary": secondary,
-        "icon_svg": icon_svg,
-        "icon_fill": icon_fill,
-        "icon_outline": icon_outline,
-        "icon_no_circle": icon_no_circle,
-        "icon_stroke_w": icon_stroke_w,
-        "letter": letter,
-    }
-
-
-class _HeadingBadgeDatum(NamedTuple):
-    """Intermediate per-badge data used to compute final positions."""
-
-    text: str
-    text_w: int
-    icon_w: int
-    total_w: int
-    show_icon: bool
-    icon_svg: markupsafe.Markup | str
-
-
-def _build_heading_context(
-    widget: Widget,
-    config: DisplayConfig,
-) -> dict[str, object]:
-    """Build Jinja2 template context for the heading widget.
-
-    Renders an optional MDI icon and heading text on a single row,
-    with optional entity badges flowing right-to-left from the right
-    edge.  Mirrors HA's Heading card (``hui-heading-card.ts``).
-
-    Heading style controls font size and fill:
-
-    - ``"title"`` — Roboto Medium at ``font_primary``, black fill
-      (default).
-    - ``"subtitle"`` — Roboto Regular at ``font_secondary``, gray fill.
-
-    Icon style controls circle rendering:
-
-    - ``"none"`` — no circle; icon glyph sized to match the heading
-      font (default).
-    - ``"filled"`` — gray-filled circle at full ``icon_dia``.
-    - ``"outlined"`` — white circle with black stroke at full
-      ``icon_dia``.
-
-    Badges are resolved right-to-left from the right edge; any badge
-    that would overlap the heading text is silently omitted.
-
-    Args:
-        widget: Widget config dict.  Recognised keys:
-            ``heading`` (display text, default ``""``),
-            ``heading_style`` (``"title"`` / ``"subtitle"``),
-            ``icon`` (MDI icon name, e.g. ``"mdi:home"``),
-            ``icon_style`` (``"none"`` / ``"filled"`` /
-            ``"outlined"``),
-            ``badges`` (list of entity ID strings or badge config
-            dicts),
-            ``card_style``, ``x``, ``w``, ``h``.
-        config: Display config with ``width``, ``states``, and
-            ``grayscale_levels``.
-
-    Returns:
-        Template context dict consumed by ``heading.svg.j2``.
-        Returns ``{"w": …, "h": …, "has_content": False,
-        **_color_context()}`` when there is nothing to render.
-        Full context includes widget dimensions, card style,
-        metrics, colors, icon geometry, heading text, and
-        pre-positioned badge list.
-    """
-    from .render import (
-        _compute_metrics,
-        _device_class_icon,
-        _load_font,
-        color_to_hex,
-    )
-
-    x = widget.get("x", PADDING)
-    svg_w = _widget_dim(widget, "w", config["width"] - x)
-    svg_h = _widget_dim(widget, "h", DEFAULT_ROW_H)
-    heading_text: str = widget.get("heading", "")
-    heading_style: str = widget.get("heading_style", "title")
-    icon_override = widget.get("icon")
-    icon_style: str = widget.get("icon_style", "none")
-    raw_badges = widget.get("badges", [])
-    card_style = widget.get("card_style", DEFAULT_CARD_STYLE)
-    states = config.get("states", {})
-    grayscale_levels = config.get("grayscale_levels", 16)
-
-    m = _compute_metrics(svg_h)
-    x_off, _r_inset, bar_width = _card_insets(m, card_style, grayscale_levels)
-
-    # Heading style: title (default) = large black; subtitle = small gray.
-    is_title = heading_style != "subtitle"
-    font_sz = m.font_primary if is_title else m.font_secondary
-    font_weight = "500" if is_title else "400"
-    colors = _color_context()
-    text_fill = colors["hex_black"] if is_title else colors["hex_gray"]
-
-    # Icon resolution: glyph size depends on circle style.
-    icon_no_circle = icon_style == "none"
-    icon_outline = icon_style == "outlined"
-    # Widen the outline stroke on 2-level displays to avoid dithering.
-    icon_stroke_w = m.border * 3 if grayscale_levels <= 2 else m.border
-    icon_svg: markupsafe.Markup | str = ""
-    glyph_sz = max(10, font_sz) if icon_no_circle else m.icon_inner
-    if icon_override is not None:
-        icon_name = str(icon_override)
-        if icon_name.startswith("mdi:"):
-            icon_name = icon_name[4:]
-        with contextlib.suppress(FileNotFoundError):
-            icon_svg = _mdi_svg_filter(icon_name, glyph_sz)
-
-    # Icon geometry: two modes depending on whether a circle is drawn.
-    icon_cx = icon_cy = icon_r = 0
-    icon_fill = ""
-    icon_glyph_x = icon_glyph_y = 0
-    content_left = x_off + m.padding
-    if icon_svg:
-        if icon_no_circle:
-            # No circle: glyph sits flush at the content left edge.
-            icon_glyph_x = content_left
-            icon_glyph_y = svg_h // 2 - glyph_sz // 2
-            text_x = content_left + glyph_sz + m.inner_gap
-        else:
-            # Circle style: glyph is centred inside the circle.
-            r = m.icon_dia // 2
-            icon_cx = content_left + r
-            icon_cy = svg_h // 2
-            icon_r = r
-            icon_fill = color_to_hex(COLOR_GRAY)
-            icon_glyph_x = icon_cx - glyph_sz // 2
-            icon_glyph_y = icon_cy - glyph_sz // 2
-            text_x = content_left + m.icon_dia + m.inner_gap
-    else:
-        text_x = content_left
-
-    text_y = svg_h // 2
-
-    # Badge font and icon sizing.
-    badge_font_sz = m.font_secondary
-    badge_font = _load_font(badge_font_sz)
-    badge_icon_sz = badge_font_sz
-
-    # Resolve badge states and compute widths.
-    badge_data: list[_HeadingBadgeDatum] = []
-    for badge_cfg in raw_badges:
-        if isinstance(badge_cfg, str):
-            entity_id: str = badge_cfg
-            show_icon = False
-            show_state = True
-            badge_icon_override = ""
-        else:
-            entity_id = badge_cfg.get("entity", "")
-            show_icon = badge_cfg.get("show_icon", False)
-            show_state = badge_cfg.get("show_state", True)
-            badge_icon_override = badge_cfg.get("icon", "")
-        state = states.get(entity_id)
-        if state is None:
-            continue
-        attrs = state.get("attributes", {})
-        state_val = state.get("state", "")
-        unit = attrs.get("unit_of_measurement", "")
-        fmtd = _fmt(state_val, config) if show_state else ""
-        badge_text = f"{fmtd}{unit}" if show_state else ""
-
-        badge_icon_svg: markupsafe.Markup | str = ""
-        if show_icon:
-            b_name: str | None = None
-            # User-supplied icon override takes priority.
-            if badge_icon_override.startswith("mdi:"):
-                b_name = badge_icon_override[4:]
-            if b_name is None:
-                domain = entity_id.split(".")[0]
-                b_name = _device_class_icon(attrs, state_val, domain)
-            if b_name is None:
-                raw = attrs.get("icon", "")
-                if raw.startswith("mdi:"):
-                    b_name = raw[4:]
-            if b_name:
-                with contextlib.suppress(FileNotFoundError):
-                    badge_icon_svg = _mdi_svg_filter(b_name, badge_icon_sz)
-
-        icon_w = (badge_icon_sz + m.inner_gap) if badge_icon_svg else 0
-        text_w = round(badge_font.getlength(badge_text))
-        badge_data.append(
-            _HeadingBadgeDatum(
-                text=badge_text,
-                text_w=text_w,
-                icon_w=icon_w,
-                total_w=icon_w + text_w,
-                show_icon=show_icon,
-                icon_svg=badge_icon_svg,
-            )
-        )
-
-    # Position badges right-to-left.  Once a badge cannot fit (its
-    # left edge would overlap the heading text), it and all remaining
-    # badges to its left in config order are dropped.
-    badge_right = svg_w - m.padding
-    rendered_badges: list[dict[str, object]] = []
-    for bd in reversed(badge_data):
-        new_right = badge_right - bd.total_w
-        if new_right < text_x + m.inner_gap:
-            break
-        badge_cy = svg_h // 2
-        rendered_badges.insert(
-            0,
-            {
-                "text": bd.text,
-                "text_x": badge_right - bd.text_w,
-                "text_y": badge_cy,
-                "show_icon": bd.show_icon,
-                "icon_svg": bd.icon_svg,
-                "icon_x": new_right,
-                "icon_y": badge_cy - badge_icon_sz // 2,
-            },
-        )
-        badge_right = new_right - m.inner_gap
-
-    has_content = bool(heading_text) or bool(icon_svg) or bool(rendered_badges)
-
-    if not has_content:
-        return {
-            "w": svg_w,
-            "h": svg_h,
-            "has_content": False,
-            **colors,
-        }
-
-    return {
-        "w": svg_w,
-        "h": svg_h,
-        "has_content": True,
-        "card_style": card_style,
-        "bar_width": bar_width,
-        **_metrics_context(m),
-        **colors,
-        # Icon geometry (only used when icon_svg is truthy).
-        "icon_svg": icon_svg,
-        "icon_cx": icon_cx,
-        "icon_cy": icon_cy,
-        "icon_r": icon_r,
-        "icon_stroke_w": icon_stroke_w,
-        "icon_fill": icon_fill,
-        "icon_color": colors["hex_black"],
-        "icon_outline": icon_outline,
-        "icon_no_circle": icon_no_circle,
-        "icon_glyph_x": icon_glyph_x,
-        "icon_glyph_y": icon_glyph_y,
-        # Heading text.
-        "heading": heading_text,
-        "font_sz": font_sz,
-        "font_weight": font_weight,
-        "text_fill": text_fill,
-        "text_x": text_x,
-        "text_y": text_y,
-        # Badges (pre-positioned, empty list when none fit).
-        "badges": rendered_badges,
-        "badge_font_sz": badge_font_sz,
-    }
-
-
 def _entity_info_context(
     widget: Widget,
     config: DisplayConfig,
@@ -2208,569 +840,20 @@ def _entity_info_context(
     }
 
 
-def _build_entity_context(
-    widget: Widget,
-    config: DisplayConfig,
-) -> dict[str, object]:
-    """Build Jinja2 template context for the entity widget.
-
-    Renders a two-section card: a header row (entity name on the
-    left, optional icon on the right) and an info section below
-    (large state value with optional unit).  Mirrors HA's Entity
-    card (``hui-entity-card.ts``).
-
-    Icon style controls circle rendering, with automatic resolution
-    based on entity state when ``icon_style`` is omitted:
-
-    - ``"filled"`` — gray-filled circle (default for active states
-      when ``grayscale_levels > 2``).
-    - ``"outlined"`` — white circle with black stroke (default for
-      inactive states and all 2-level displays).
-    - ``"none"`` — no circle; icon glyph rendered without decoration.
-
-    Args:
-        widget: Widget config dict.  Recognised keys:
-            ``entity`` (HA entity ID, required),
-            ``name`` (display name override),
-            ``icon`` (MDI icon name, e.g. ``"mdi:thermometer"``),
-            ``attribute`` (attribute key to show as value instead of
-            state),
-            ``unit`` (unit string override),
-            ``icon_style`` (``"filled"`` / ``"outlined"`` /
-            ``"none"``),
-            ``card_style``, ``x``, ``w``, ``h``.
-        config: Display config with ``width``, ``states``, and
-            ``grayscale_levels``.
-
-    Returns:
-        Template context dict consumed by ``entity.svg.j2``.
-        Returns ``{"w": …, "h": …, "has_entity": False,
-        **_color_context()}`` when the entity is missing.
-        Full context includes widget dimensions, card style,
-        metrics, colors, icon geometry, header text, and info
-        section value/unit.
-    """
-    x = widget.get("x", PADDING)
-    svg_w = _widget_dim(widget, "w", config["width"] - x)
-    svg_h = _widget_dim(widget, "h", 2 * DEFAULT_ROW_H)
-    attribute: str | None = widget.get("attribute")
-    ctx = _entity_info_context(
-        widget, config, svg_h, svg_w, svg_h, attribute=attribute
-    )
-    if ctx is None:
-        return {
-            "w": svg_w,
-            "h": svg_h,
-            "has_entity": False,
-            **_color_context(),
-        }
-    return ctx
-
-
-def _build_entities_context(
-    widget: Widget,
-    config: DisplayConfig,
-) -> dict[str, object]:
-    """Build Jinja2 template context for the entities widget.
-
-    Multi-entity list card with optional title and per-row types.
-    Mirrors HA's Entities card (``hui-entities-card.ts``).
-
-    Each item in the ``entities`` config list is classified as one
-    of three row types:
-
-    - **Entity row** — a plain string entity ID, or a dict with an
-      ``entity`` key (and optional ``name`` / ``icon`` overrides).
-      Rendered with icon circle, primary text, and right-aligned
-      state value.
-    - **Divider row** — ``{"type": "divider"}``.  Rendered as a
-      thin gray horizontal line.
-    - **Section row** — ``{"type": "section", "label": str}``.
-      Rendered as a gray sub-heading above the next group.
-
-    Entity rows whose entity ID is absent from ``states`` are
-    silently skipped.  The widget returns ``has_rows=False`` when
-    no entity rows remain after filtering.
-
-    Icon style is widget-level and applies uniformly to all entity
-    rows:
-
-    - ``"filled"`` — gray-filled circle, white icon (default for
-      active states on multi-level displays).
-    - ``"outlined"`` — white fill, black stroke circle, black icon
-      (default for inactive states and 2-level displays).
-    - ``"none"`` — no circle; icon glyph rendered in black.
-
-    Auto-dividers — thin gray lines drawn between consecutive entity
-    rows — are suppressed when a divider or section row follows the
-    entity row, since those rows provide their own visual break.
-
-    Args:
-        widget: Widget config dict.  Recognised keys:
-            ``entities`` (list of row configs),
-            ``title`` (optional label above the card),
-            ``icon_style`` (``"filled"`` / ``"outlined"`` /
-            ``"none"``), ``card_style``, ``x``, ``w``, ``h``.
-        config: Display config with ``width``, ``states``, and
-            ``grayscale_levels``.
-
-    Returns:
-        Template context dict consumed by ``entities.svg.j2``.
-        Returns ``{"w": …, "h": …, "has_rows": False,
-        **_color_context()}`` when no entity rows survive
-        state filtering.  Full context includes ``w``, ``h``,
-        ``has_rows``, title layout fields, card geometry,
-        metrics (``m_*``), colors (``hex_*``), row list and
-        per-row layout dicts.
-    """
-    from .render import (
-        _compute_metrics,
-        _device_class_icon,
-        color_to_hex,
-    )
-
-    x = widget.get("x", PADDING)
-    svg_w = _widget_dim(widget, "w", config["width"] - x)
-    title: str = widget.get("title", "")
-    icon_style = widget.get("icon_style")
-    card_style = widget.get("card_style", DEFAULT_CARD_STYLE)
-    entity_configs: list = widget.get("entities", [])
-    states = config.get("states", {})
-    grayscale_levels = config.get("grayscale_levels", 16)
-    colors = _color_context()
-
-    # --- Classify rows ---
-    # Absent entity states are filtered out here so layout counts
-    # reflect what will actually be rendered.
-    classified: list[dict[str, object]] = []
-    for entry in entity_configs:
-        if isinstance(entry, str):
-            if entry in states:
-                classified.append(
-                    {
-                        "row_type": "entity",
-                        "entity_id": entry,
-                        "name_override": None,
-                        "icon_override": None,
-                    }
-                )
-        elif isinstance(entry, dict):
-            row_type = entry.get("type")
-            if row_type == "divider":
-                classified.append({"row_type": "divider"})
-            elif row_type == "section":
-                classified.append(
-                    {
-                        "row_type": "section",
-                        "label": str(entry.get("label", "")),
-                    }
-                )
-            else:
-                eid: str = str(entry.get("entity", ""))
-                if eid and eid in states:
-                    classified.append(
-                        {
-                            "row_type": "entity",
-                            "entity_id": eid,
-                            "name_override": entry.get("name"),
-                            "icon_override": entry.get("icon"),
-                        }
-                    )
-
-    n_entity = sum(1 for r in classified if r["row_type"] == "entity")
-    if n_entity == 0:
-        return {
-            "w": svg_w,
-            "h": _widget_dim(widget, "h", DEFAULT_ROW_H),
-            "has_rows": False,
-            **colors,
-        }
-
-    n_divider = sum(1 for r in classified if r["row_type"] == "divider")
-    n_section = sum(1 for r in classified if r["row_type"] == "section")
-
-    # Divider takes 15% of an entity row height.
-    divider_h_base = round(DEFAULT_ROW_H * 0.15)
-    # Section sub-heading takes 60% of an entity row height.
-    section_h_base = round(DEFAULT_ROW_H * 0.6)
-
-    # --- Height ---
-    if "h" not in widget:
-        # Auto-height: delegate to the shared fixpoint helper,
-        # passing a heterogeneous content_target that accounts for
-        # the mix of entity, divider, and section row heights.
-        content_target = (
-            n_entity * DEFAULT_ROW_H
-            + n_divider * divider_h_base
-            + n_section * section_h_base
-        )
-        svg_h = _auto_row_height(
-            title,
-            max(1, n_entity),
-            content_target=content_target,
-        )
-        title_font_sz, content_y, content_h = _title_layout(title, svg_h)
-        row_h = DEFAULT_ROW_H
-    else:
-        svg_h = _widget_dim(widget, "h", DEFAULT_ROW_H)
-        title_font_sz, content_y, content_h = _title_layout(title, svg_h)
-        # Proportional allocation: entity rows have weight 1,
-        # divider rows 0.15, section rows 0.6.
-        total_weight = n_entity + n_divider * 0.15 + n_section * 0.6
-        row_h = max(1, round(content_h / total_weight))
-
-    # Sub-row heights scale with row_h.
-    # Divider takes 15% of an entity row height.
-    divider_h = max(1, round(row_h * 0.15))
-    # Section sub-heading takes 60% of an entity row height.
-    section_h = max(1, round(row_h * 0.6))
-
-    m = _compute_metrics(row_h)
-    x_off, r_inset, bar_width = _card_insets(m, card_style, grayscale_levels)
-    lpad = m.padding if x_off == 0 else 0
-    rpad = m.padding if r_inset == 0 else 0
-
-    # Widen outline stroke on 2-level displays to avoid dithering.
-    icon_stroke_w = m.border * 3 if grayscale_levels <= 2 else m.border
-    # Widen divider lines on 2-level displays (same rationale).
-    divider_stroke_w = m.divider * 3 if grayscale_levels <= 2 else m.divider
-    icon_fill = color_to_hex(COLOR_GRAY)
-    # Section label font ~46% of row height (matches chip label ratio).
-    section_font_sz = round(row_h * 0.46)
-
-    # --- Build row dicts ---
-    rows: list[dict[str, object]] = []
-    y = content_y  # absolute SVG y pointer within the canvas
-
-    for i, cls in enumerate(classified):
-        if cls["row_type"] == "divider":
-            rows.append(
-                {
-                    "row_type": "divider",
-                    "line_y": y + divider_h // 2,
-                }
-            )
-            y += divider_h
-
-        elif cls["row_type"] == "section":
-            rows.append(
-                {
-                    "row_type": "section",
-                    "label": str(cls.get("label", "")),
-                    "font_sz": section_font_sz,
-                    "text_y": y + section_h // 2,
-                }
-            )
-            y += section_h
-
-        else:
-            entity_id = str(cls["entity_id"])
-            name_override = cls["name_override"]
-            icon_override = cls["icon_override"]
-            state = states[entity_id]
-            attrs = state.get("attributes", {})
-            domain = entity_id.split(".")[0]
-            state_val: str = state.get("state", "")
-
-            primary: str = (
-                str(name_override)
-                if name_override is not None
-                else attrs.get("friendly_name", entity_id)
-            )
-
-            unit = attrs.get("unit_of_measurement", "")
-            fmtd = _fmt(state_val, config)
-            value = f"{fmtd}{unit}" if unit else fmtd
-
-            # Icon: explicit override → device_class →
-            # attrs["icon"] → letter fallback.
-            icon_svg: markupsafe.Markup | str = ""
-            if icon_override is not None:
-                icon_name: str | None = str(icon_override)
-                if icon_name.startswith("mdi:"):
-                    icon_name = icon_name[4:]
-                with contextlib.suppress(FileNotFoundError):
-                    icon_svg = _mdi_svg_filter(icon_name, m.icon_inner)
-            else:
-                icon_name = _device_class_icon(attrs, state_val, domain)
-                if icon_name is None:
-                    raw = attrs.get("icon", "")
-                    if raw.startswith("mdi:"):
-                        icon_name = raw[4:]
-                if icon_name:
-                    with contextlib.suppress(FileNotFoundError):
-                        icon_svg = _mdi_svg_filter(icon_name, m.icon_inner)
-
-            letter = ""
-            if not icon_svg:
-                friendly = attrs.get("friendly_name", entity_id)
-                letter = friendly[:1].upper() if friendly else ""
-
-            # Icon style: widget-level or auto by entity state.
-            is_active = state_val in _ACTIVE_STATES
-            if icon_style is None:
-                resolved_style = (
-                    "outlined"
-                    if grayscale_levels <= 2
-                    else ("filled" if is_active else "outlined")
-                )
-            else:
-                resolved_style = str(icon_style)
-
-            icon_outline = resolved_style == "outlined"
-            icon_no_circle = resolved_style == "none"
-
-            # Auto-divider: draw after this entity row only when
-            # the immediately following row is also an entity row.
-            # Explicit divider/section rows break the chain.
-            draw_divider = (
-                i + 1 < len(classified)
-                and classified[i + 1]["row_type"] == "entity"
-            )
-
-            # Clamp so rounding in the proportional-height branch
-            # cannot push the last row past the content area bottom.
-            row_y = min(y, content_y + content_h - row_h)
-            rows.append(
-                {
-                    "row_type": "entity",
-                    "y": row_y,
-                    "primary": primary,
-                    "value": value,
-                    "icon_svg": icon_svg,
-                    "icon_fill": icon_fill,
-                    "icon_outline": icon_outline,
-                    "icon_no_circle": icon_no_circle,
-                    "letter": letter,
-                    "draw_divider": draw_divider,
-                }
-            )
-            y += row_h
-
-    return {
-        "w": svg_w,
-        "h": svg_h,
-        "has_rows": True,
-        "title": title,
-        "title_font_sz": title_font_sz,
-        "content_y": content_y,
-        "content_h": content_h,
-        "card_style": card_style,
-        "bar_width": bar_width,
-        **_metrics_context(m),
-        **colors,
-        "row_h": row_h,
-        "rows": rows,
-        "x_off": x_off,
-        "r_inset": r_inset,
-        "lpad": lpad,
-        "rpad": rpad,
-        "icon_stroke_w": icon_stroke_w,
-        "divider_stroke_w": divider_stroke_w,
-    }
-
-
-def _build_sensor_context(
-    widget: Widget,
-    config: DisplayConfig,
-) -> dict[str, object]:
-    """Build Jinja2 template context for the sensor widget.
-
-    Renders a header row (entity name on the left, optional icon on
-    the right), an info section (large state value with unit), and an
-    optional history sparkline graph below.  Mirrors HA's Sensor card.
-
-    Icon style follows the same auto-resolution logic as the Entity
-    widget: active states → filled circle; inactive or 2-level
-    displays → outlined circle; ``icon_style="none"`` hides the
-    circle entirely.
-
-    Without ``graph="line"``, the default height is
-    ``2 * DEFAULT_ROW_H``.  With ``graph="line"``, the default is
-    ``3 * DEFAULT_ROW_H`` (entity section + graph row).  History
-    data is read from ``states[entity_id]["history"]``, injected by
-    ``_fetch_history()`` in ``__init__.py``.
-
-    Args:
-        widget: Widget config dict.  Recognised keys:
-            ``entity`` (HA entity ID, required),
-            ``name`` (display name override),
-            ``icon`` (MDI icon name, e.g. ``"mdi:thermometer"``),
-            ``unit`` (unit string override),
-            ``icon_style`` (``"filled"`` / ``"outlined"`` /
-            ``"none"``),
-            ``graph`` (``"line"`` to enable history sparkline),
-            ``hours_to_show`` (history window in hours; default 24),
-            ``detail`` (``1`` for downsampled ~24 pts, ``2`` for
-            full resolution; default 1),
-            ``limits`` (dict with optional ``"min"``/``"max"``
-            keys to fix the Y-axis range; editor uses flat
-            ``limits_min``/``limits_max`` keys instead),
-            ``card_style``, ``x``, ``w``, ``h``.
-        config: Display config with ``width``, ``states``, and
-            ``grayscale_levels``.
-
-    Returns:
-        Template context dict consumed by ``sensor.svg.j2``.
-        Returns ``{"w": …, "h": …, "has_entity": False,
-        **_color_context()}`` when the entity is missing.
-        Full context includes widget dimensions, card style,
-        metrics, colors, icon geometry, header text, info section
-        value/unit, and graph polyline/fill data.
-    """
-    from .render import DEFAULT_METRICS
-
-    x = widget.get("x", PADDING)
-    svg_w = _widget_dim(widget, "w", config["width"] - x)
-    graph: str | None = widget.get("graph")
-    detail: int = int(float(widget.get("detail", 1)))
-    hours_to_show: int = int(float(widget.get("hours_to_show", 24)))
-    limits: dict[str, float] | None = widget.get("limits")
-    if limits is None:
-        # The frontend stores Y-axis bounds as flat keys rather than a
-        # nested dict (ha-form does not support nested objects).
-        lmin = widget.get("limits_min")
-        lmax = widget.get("limits_max")
-        if lmin is not None or lmax is not None:
-            limits = {}
-            if lmin is not None:
-                limits["min"] = lmin
-            if lmax is not None:
-                limits["max"] = lmax
-    grayscale_levels = config.get("grayscale_levels", 16)
-
-    has_graph = graph == "line"
-    # Default height: 3 rows with graph section, 2 rows without.
-    default_h = 3 * DEFAULT_ROW_H if has_graph else 2 * DEFAULT_ROW_H
-    svg_h = _widget_dim(widget, "h", default_h)
-    # Graph occupies the bottom DEFAULT_ROW_H pixels; entity section
-    # fills the rest.
-    graph_h = DEFAULT_ROW_H if has_graph else 0
-    entity_h = max(1, svg_h - graph_h)
-
-    ctx = _entity_info_context(widget, config, entity_h, svg_w, svg_h)
-    if ctx is None:
-        return {
-            "w": svg_w,
-            "h": svg_h,
-            "has_entity": False,
-            **_color_context(),
-        }
-
-    # name_x = x_off + lpad (left content edge); icon_cx + icon_r =
-    # svg_w - r_inset - rpad (right content edge without icon overlap).
-    gx1: int = cast("int", ctx["name_x"])
-    gx2: int = cast("int", ctx["icon_cx"]) + cast("int", ctx["icon_r"])
-
-    # --- Graph section ---
-    polyline_points = ""
-    fill_points = ""
-    # Graph stroke uses DEFAULT_METRICS (fixed at DEFAULT_ROW_H) because
-    # the graph zone is always DEFAULT_ROW_H pixels tall regardless of
-    # entity section height.  Widened 2× on 2-level displays.
-    graph_stroke_w = (
-        DEFAULT_METRICS.border * 2
-        if grayscale_levels <= 2
-        else DEFAULT_METRICS.border
-    )
-
-    if has_graph:
-        entity_id: str = widget.get("entity", "")
-        state = config.get("states", {}).get(entity_id, {})
-        history: list[dict[str, object]] = state.get("history", [])
-        # Apply hours_to_show window: keep only entries whose timestamp
-        # falls within the most recent hours_to_show hours.
-        if history:
-            t_latest = max(float(str(e.get("lu", 0))) for e in history)
-            cutoff = t_latest - hours_to_show * 3600
-            history = [
-                e for e in history if float(str(e.get("lu", 0))) > cutoff
-            ]
-        # Filter entries to numeric values only; skip unavailable /
-        # unknown / on/off states so only plottable data remains.
-        numeric: list[tuple[float, float]] = []
-        for entry in history:
-            s = entry.get("s", "")
-            lu = entry.get("lu", 0.0)
-            try:
-                val = float(str(s))
-                numeric.append((float(str(lu)), val))
-            except (ValueError, TypeError):
-                continue
-
-        if len(numeric) >= 2:
-            # Downsample to ~24 points for detail=1 to avoid
-            # pixel-level noise on small sparklines.
-            if detail <= 1 and len(numeric) > 24:
-                step = len(numeric) / 24
-                indices = {round(i * step) for i in range(24)}
-                indices.add(len(numeric) - 1)
-                numeric = [
-                    numeric[i] for i in sorted(indices) if i < len(numeric)
-                ]
-
-            # Y range: explicit limits override auto min/max.
-            # isinstance guard rejects null or string values gracefully.
-            values = [v for _, v in numeric]
-            y_min = (
-                float(limits["min"])
-                if limits
-                and "min" in limits
-                and isinstance(limits["min"], (int, float))
-                else min(values)
-            )
-            y_max = (
-                float(limits["max"])
-                if limits
-                and "max" in limits
-                and isinstance(limits["max"], (int, float))
-                else max(values)
-            )
-            if y_max == y_min:
-                # Flat line: add a unit pad so the polyline is
-                # centred rather than collapsed to the bottom edge.
-                y_max += 1.0
-
-            # Inset by 2× stroke width so the polyline stays
-            # inside the widget bounds (wider on 2-level displays).
-            margin = graph_stroke_w * 2
-            gy1 = entity_h + margin
-            gy2 = svg_h - margin
-
-            timestamps = [t for t, _ in numeric]
-            t_min = min(timestamps)
-            t_range = max(max(timestamps) - t_min, 1.0)
-            y_range = y_max - y_min
-
-            # Map (timestamp, value) → (x, y) pixel; Y is inverted.
-            pts = [
-                (
-                    round(gx1 + (t - t_min) / t_range * (gx2 - gx1)),
-                    round(gy2 - (v - y_min) / y_range * (gy2 - gy1)),
-                )
-                for t, v in numeric
-            ]
-            polyline_points = " ".join(f"{px},{py}" for px, py in pts)
-            # Fill polygon rendered on all display types; on 2-level
-            # displays Floyd-Steinberg dithering in optimize.py
-            # converts the light-gray fill to a dot pattern.
-            bottom = gy2
-            fill_pts = [
-                *pts,
-                (pts[-1][0], bottom),
-                (pts[0][0], bottom),
-            ]
-            fill_points = " ".join(f"{px},{py}" for px, py in fill_pts)
-
-    return {
-        **ctx,
-        # Graph.
-        "has_graph": has_graph,
-        "polyline_points": polyline_points,
-        "fill_points": fill_points,
-        "graph_stroke_w": graph_stroke_w,
-    }
-
+# Import widget builders at module bottom to avoid circular imports:
+# widget modules import helpers defined above; by this point all
+# helpers exist in the partially-loaded svg_render module namespace.
+from .widgets import (  # noqa: E402
+    _build_device_battery_context,
+    _build_entities_context,
+    _build_entity_context,
+    _build_heading_context,
+    _build_sensor_context,
+    _build_separator_context,
+    _build_tile_context,
+    _build_waste_schedule_context,
+    _build_weather_context,
+)
 
 _SVG_RENDERERS: dict[str, SvgContextFn] = {
     WidgetType.DEVICE_BATTERY: _build_device_battery_context,
